backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.5
pydantic==2.10.3
Pillow==11.0.0
